Minor Linux bugfix
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2734 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_compumotor.c
@@ -385,7 +385,7 @@ mxd_compumotor_check_for_servo( MX_COMPUMOTOR_INTERFACE *compumotor_interface,
 
 #if MXD_COMPUMOTOR_DEBUG
 		MX_DEBUG(-2,("%s: response[%d] = '%c'",
-				fname, l, response[l]));
+				fname, (int) l, response[l]));
 #endif
 
 		if ( response[l] == '1' ) {
